Implemented tests for the various cleanup algorithms.
We also don't delete items that are in Favorites.
