+++ app/src/androidTest/java/de/test/antennapod/storage/DBQueueCleanupAlgorithmTest.java
@@ -1,10 +1,6 @@
 package de.test.antennapod.storage;
 
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
 import android.test.FlakyTest;
-import android.test.InstrumentationTestCase;
 
 import java.io.File;
 import java.io.IOException;
@@ -19,67 +15,23 @@ import de.danoeh.antennapod.core.preferences.UserPreferences;
 import de.danoeh.antennapod.core.storage.DBTasks;
 import de.danoeh.antennapod.core.storage.PodDBAdapter;
 
-import static de.test.antennapod.storage.DBTestUtils.saveFeedlist;
-
 /**
  * Tests that the APQueueCleanupAlgorithm is working correctly.
  */
-public class DBQueueCleanupAlgorithmTest extends InstrumentationTestCase {
+public class DBQueueCleanupAlgorithmTest extends DBCleanupTests {
 
     private static final String TAG = "DBQueueCleanupAlgorithmTest";
-    private static final int EPISODE_CACHE_SIZE = 5;
 
-    private Context context;
-    
-    private File destFolder;
-
-    @Override
-    protected void tearDown() throws Exception {
-        super.tearDown();
-
-        assertTrue(PodDBAdapter.deleteDatabase());
-
-        cleanupDestFolder(destFolder);
-        assertTrue(destFolder.delete());
-    }
-
-    private void cleanupDestFolder(File destFolder) {
-        for (File f : destFolder.listFiles()) {
-            assertTrue(f.delete());
-        }
-    }
-
-    @Override
-    protected void setUp() throws Exception {
-        super.setUp();
-        context = getInstrumentation().getTargetContext();
-        destFolder = context.getExternalCacheDir();
-        cleanupDestFolder(destFolder);
-        assertNotNull(destFolder);
-        assertTrue(destFolder.exists());
-        assertTrue(destFolder.canWrite());
-
-        // create new database
-        PodDBAdapter.deleteDatabase();
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.close();
-
-        SharedPreferences.Editor prefEdit = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext()).edit();
-        prefEdit.putString(UserPreferences.PREF_EPISODE_CACHE_SIZE, Integer.toString(EPISODE_CACHE_SIZE));
-        prefEdit.putInt(UserPreferences.PREF_EPISODE_CLEANUP, UserPreferences.EPISODE_CLEANUP_QUEUE);
-        prefEdit.commit();
-
-        UserPreferences.init(context);
+    public DBQueueCleanupAlgorithmTest() {
+        super(UserPreferences.EPISODE_CLEANUP_QUEUE);
     }
 
     /**
-     * A test with no items in the queue, but multiple items downloaded. A certain
-     * number of items should be removed to make room in in the cache/queue.
-     * @throws IOException
+     * For APQueueCleanupAlgorithm we expect even unplayed episodes to be deleted if needed
+     * if they aren't in the queue
      */
     @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldDelete() throws IOException {
+    public void testPerformAutoCleanupHandleUnplayed() throws IOException {
         final int NUM_ITEMS = EPISODE_CACHE_SIZE * 2;
 
         Feed feed = new Feed("url", new Date(), "title");
@@ -114,64 +66,5 @@ public class DBQueueCleanupAlgorithmTest extends InstrumentationTestCase {
                 assertFalse(files.get(i).exists());
             }
         }
-    }
-
-    @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldNotDeleteBecauseInQueue() throws IOException {
-        final int NUM_ITEMS = EPISODE_CACHE_SIZE * 2;
-
-        Feed feed = new Feed("url", new Date(), "title");
-        List<FeedItem> items = new ArrayList<>();
-        feed.setItems(items);
-        List<File> files = new ArrayList<>();
-        for (int i = 0; i < NUM_ITEMS; i++) {
-            FeedItem item = new FeedItem(0, "title", "id", "link", new Date(NUM_ITEMS - i), FeedItem.PLAYED, feed);
-
-            File f = new File(destFolder, "file " + i);
-            assertTrue(f.createNewFile());
-            assertTrue(f.exists());
-            files.add(f);
-            item.setMedia(new FeedMedia(0, item, 1, 0, 1L, "m", f.getAbsolutePath(), "url", true, new Date(NUM_ITEMS - i), 0));
-            items.add(item);
-        }
-
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.setCompleteFeed(feed);
-        adapter.setQueue(items);
-        adapter.close();
-
-        assertTrue(feed.getId() != 0);
-        for (FeedItem item : items) {
-            assertTrue(item.getId() != 0);
-            assertTrue(item.getMedia().getId() != 0);
-        }
-        DBTasks.performAutoCleanup(context);
-        for (File file : files) {
-            assertTrue(file.exists());
-        }
-    }
-
-    /**
-     * Reproduces a bug where DBTasks.performAutoCleanup(android.content.Context) would use the ID of the FeedItem in the
-     * call to DBWriter.deleteFeedMediaOfItem instead of the ID of the FeedMedia. This would cause the wrong item to be deleted.
-     * @throws IOException
-     */
-    @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldNotDeleteBecauseInQueue_withFeedsWithNoMedia() throws IOException {
-        // add feed with no enclosures so that item ID != media ID
-        saveFeedlist(1, 10, false);
-
-        // add candidate for performAutoCleanup
-        List<Feed> feeds = saveFeedlist(1, 1, true);
-        FeedMedia m = feeds.get(0).getItems().get(0).getMedia();
-        m.setDownloaded(true);
-        m.setFile_url("file");
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.setMedia(m);
-        adapter.close();
-
-        testPerformAutoCleanupShouldNotDeleteBecauseInQueue();
     }
 }
