+++ app/src/androidTest/java/de/test/antennapod/storage/DBTasksTest.java
@@ -29,33 +29,20 @@ import static de.test.antennapod.storage.DBTestUtils.saveFeedlist;
 public class DBTasksTest extends InstrumentationTestCase {
 
     private static final String TAG = "DBTasksTest";
-    private static final int EPISODE_CACHE_SIZE = 5;
 
     private Context context;
-    
-    private File destFolder;
 
     @Override
     protected void tearDown() throws Exception {
         super.tearDown();
 
         assertTrue(PodDBAdapter.deleteDatabase());
-
-        for (File f : destFolder.listFiles()) {
-            assertTrue(f.delete());
-        }
-        assertTrue(destFolder.delete());
-
     }
 
     @Override
     protected void setUp() throws Exception {
         super.setUp();
         context = getInstrumentation().getTargetContext();
-        destFolder = context.getExternalCacheDir();
-        assertNotNull(destFolder);
-        assertTrue(destFolder.exists());
-        assertTrue(destFolder.canWrite());
 
         // create new database
         PodDBAdapter.deleteDatabase();
@@ -63,144 +50,7 @@ public class DBTasksTest extends InstrumentationTestCase {
         adapter.open();
         adapter.close();
 
-        SharedPreferences.Editor prefEdit = PreferenceManager.getDefaultSharedPreferences(context.getApplicationContext()).edit();
-        prefEdit.putString(UserPreferences.PREF_EPISODE_CACHE_SIZE, Integer.toString(EPISODE_CACHE_SIZE));
-        prefEdit.putInt(UserPreferences.PREF_EPISODE_CLEANUP, UserPreferences.EPISODE_CLEANUP_DEFAULT);
-        prefEdit.commit();
-
         UserPreferences.init(context);
-    }
-
-    @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldDelete() throws IOException {
-        final int NUM_ITEMS = EPISODE_CACHE_SIZE * 2;
-
-        Feed feed = new Feed("url", new Date(), "title");
-        List<FeedItem> items = new ArrayList<>();
-        feed.setItems(items);
-        List<File> files = new ArrayList<>();
-        for (int i = 0; i < NUM_ITEMS; i++) {
-            FeedItem item = new FeedItem(0, "title", "id", "link", new Date(), FeedItem.PLAYED, feed);
-
-            File f = new File(destFolder, "file " + i);
-            assertTrue(f.createNewFile());
-            files.add(f);
-            item.setMedia(new FeedMedia(0, item, 1, 0, 1L, "m", f.getAbsolutePath(), "url", true, new Date(NUM_ITEMS - i), 0));
-            items.add(item);
-        }
-
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.setCompleteFeed(feed);
-        adapter.close();
-
-        assertTrue(feed.getId() != 0);
-        for (FeedItem item : items) {
-            assertTrue(item.getId() != 0);
-            assertTrue(item.getMedia().getId() != 0);
-        }
-        DBTasks.performAutoCleanup(context);
-        for (int i = 0; i < files.size(); i++) {
-            if (i < EPISODE_CACHE_SIZE) {
-                assertTrue(files.get(i).exists());
-            } else {
-                assertFalse(files.get(i).exists());
-            }
-        }
-    }
-
-    @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldNotDeleteBecauseUnplayed() throws IOException {
-        final int NUM_ITEMS = EPISODE_CACHE_SIZE * 2;
-
-        Feed feed = new Feed("url", new Date(), "title");
-        List<FeedItem> items = new ArrayList<FeedItem>();
-        feed.setItems(items);
-        List<File> files = new ArrayList<File>();
-        for (int i = 0; i < NUM_ITEMS; i++) {
-            FeedItem item = new FeedItem(0, "title", "id", "link", new Date(), FeedItem.UNPLAYED, feed);
-
-            File f = new File(destFolder, "file " + i);
-            assertTrue(f.createNewFile());
-            assertTrue(f.exists());
-            files.add(f);
-            item.setMedia(new FeedMedia(0, item, 1, 0, 1L, "m", f.getAbsolutePath(), "url", true, new Date(NUM_ITEMS - i), 0));
-            items.add(item);
-        }
-
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.setCompleteFeed(feed);
-        adapter.close();
-
-        assertTrue(feed.getId() != 0);
-        for (FeedItem item : items) {
-            assertTrue(item.getId() != 0);
-            assertTrue(item.getMedia().getId() != 0);
-        }
-        DBTasks.performAutoCleanup(context);
-        for (File file : files) {
-            assertTrue(file.exists());
-        }
-    }
-
-    @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldNotDeleteBecauseInQueue() throws IOException {
-        final int NUM_ITEMS = EPISODE_CACHE_SIZE * 2;
-
-        Feed feed = new Feed("url", new Date(), "title");
-        List<FeedItem> items = new ArrayList<>();
-        feed.setItems(items);
-        List<File> files = new ArrayList<>();
-        for (int i = 0; i < NUM_ITEMS; i++) {
-            FeedItem item = new FeedItem(0, "title", "id", "link", new Date(), FeedItem.PLAYED, feed);
-
-            File f = new File(destFolder, "file " + i);
-            assertTrue(f.createNewFile());
-            assertTrue(f.exists());
-            files.add(f);
-            item.setMedia(new FeedMedia(0, item, 1, 0, 1L, "m", f.getAbsolutePath(), "url", true, new Date(NUM_ITEMS - i), 0));
-            items.add(item);
-        }
-
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.setCompleteFeed(feed);
-        adapter.setQueue(items);
-        adapter.close();
-
-        assertTrue(feed.getId() != 0);
-        for (FeedItem item : items) {
-            assertTrue(item.getId() != 0);
-            assertTrue(item.getMedia().getId() != 0);
-        }
-        DBTasks.performAutoCleanup(context);
-        for (File file : files) {
-            assertTrue(file.exists());
-        }
-    }
-
-    /**
-     * Reproduces a bug where DBTasks.performAutoCleanup(android.content.Context) would use the ID of the FeedItem in the
-     * call to DBWriter.deleteFeedMediaOfItem instead of the ID of the FeedMedia. This would cause the wrong item to be deleted.
-     * @throws IOException
-     */
-    @FlakyTest(tolerance = 3)
-    public void testPerformAutoCleanupShouldNotDeleteBecauseInQueue_withFeedsWithNoMedia() throws IOException {
-        // add feed with no enclosures so that item ID != media ID
-        saveFeedlist(1, 10, false);
-
-        // add candidate for performAutoCleanup
-        List<Feed> feeds = saveFeedlist(1, 1, true);
-        FeedMedia m = feeds.get(0).getItems().get(0).getMedia();
-        m.setDownloaded(true);
-        m.setFile_url("file");
-        PodDBAdapter adapter = PodDBAdapter.getInstance();
-        adapter.open();
-        adapter.setMedia(m);
-        adapter.close();
-
-        testPerformAutoCleanupShouldNotDeleteBecauseInQueue();
     }
 
     @FlakyTest(tolerance = 3)
